Included the Visual Studio project
Since this a whole application and not just a utiility which can just include files, I include the visual studio project. This will build to an executable that will be able to be run when built. It also allows me to include a build folder in order to maintain a project structure.