Currently working on reworking the rendering architecture.
I'd like the interface to be as simple as possible and leave as much details to the implementation as possible. This allows me to do a lot of cool things with the materials depending on the renderer implementation. I want to be able to construct the materials in place instead of building them up every time I run the engine. This is alright because I'm not really aiming for cross platform support.